Patch the actual numpy import, not just the module
